+++ changelog.md
@@ -3,6 +3,7 @@
 
 
  - [feature] storing user's preferred language in user settings
+ - [feature] PRT-74 update renamed repositories
  - [feature] PR-closed-notification: backend part (w/o WebSocket support)
  - [feature] session timeout 9 hrs
  - [feature] storing build uri in pullrequest when fetching pullrequest status from Github
@@ -52,4 +53,3 @@
  - [initial] initial commit
 
 -- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed Apr 22 12:17:41 CEST 2015
-